chore(testings): updating testings and fully coverage
