+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "@platformbuilders/validations",
-  "version": "0.1.3",
-  "description": "",
-  "author": "Platform Builders <apps@platformbuilders.io>",
+  "version": "0.1.4",
+  "description": "Validações de comuns de formulários",
+  "author": "Platform Builders <digitals@platformbuilders.io>",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/platformbuilders/validations.git"
   },
   "keywords": [
     "platformbuilders",
+    "builders",
     "validations"
   ],
   "license": "ISC",
@@ -21,7 +22,7 @@
     "npm": ">=6"
   },
   "scripts": {
-    "test": "jest --passWithNoTests",
+    "test": "jest --passWithNoTests --coverage",
     "test:watch": "react-scripts test --env=jsdom",
     "build": "rm -rf dist && rollup -c --bundleConfigAsCjs && yarn uglify",
     "start": "rollup -c -w --bundleConfigAsCjs",
@@ -30,6 +31,13 @@
     "uglify": "for f in dist/*.js; do short=${f%.js}; uglifyjs $f > $short.min.js; rm -rf $f; mv $short.min.js $short.js; done",
     "lint": "tsc && eslint '*/**/*.{ts, tsx}' --fix"
   },
+    "lint-staged": {
+      "*.{ts,tsx}": [
+        "yarn pretty",
+        "yarn lint",
+        "yarn test"
+      ]
+    },
   "peerDependencies": {
     "@fnando/cnpj": "1.0.2",
     "@fnando/cpf": "1.0.2",
