+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './isValidCnpj';
 export * from './isValidCep';
 export * from './isValidCpf';
 export * from './isValidCurrency';
+export * from './isValidDDD';
 export * from './isValidPhone';
 export * from './isValidDocument';
 export * from './isValidEmail';
@@ -16,4 +17,5 @@ export * from './isValidAccountBank';
 export * from './isValidAlphanumeric';
 export * from './isValidOnlyNumbers';
 export * from './isValidAlphabetLetters';
-export * from './isValidInstagramUserName';
+export * from './isValidSocialUsername';
+export * from './isValidNickName';
