+++ src/isValidSocialName.ts
@@ -1,8 +1,3 @@
-export const isValidSocialName = (socialName: string): boolean =>
-  !!socialName && socialName.length >= 2 && socialName.length <= 10;
-
-export const isValidNickName = (nickName: string): boolean =>
-  !!nickName && nickName.length >= 3 && nickName.length <= 10;
 /**
  * Validates a social name and the criteria to approve a social is:
  * * Must have only letters
@@ -11,5 +6,5 @@ export const isValidNickName = (nickName: string): boolean =>
  * @param socialName
  * @returns boolean
  */
-export const isValidSocialNameV2 = (socialName?: string): boolean =>
+export const isValidSocialName = (socialName?: string): boolean =>
   socialName ? /^[a-zA-Z][a-zA-Z\s]+$/.test(socialName) : false;
